# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Qui Browser API",
    "description": "Comprehensive API for Qui Browser - A lightweight, secure web browser server with VR/WebXR support",
    "version": "1.1.0",
    "contact": {
      "name": "Qui Browser Team",
      "url": "https://github.com/qui-browser",
      "email": "team@qui-browser.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.qui-browser.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
  "paths": {
...
    "/api/bookmarks": {
      "get": {
        "summary": "Get Bookmarks",
        "description": "Retrieve all bookmarks for the authenticated user",
        "tags": ["Bookmarks"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "List of bookmarks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Bookmark"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Add Bookmark",
        "description": "Create a new bookmark",
        "tags": ["Bookmarks"],
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bookmark created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bookmark"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/history": {
      "get": {
        "summary": "Get Browsing History",
        "description": "Retrieve browsing history with optional filters",
        "tags": ["History"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            },
            "description": "Maximum number of items to return"
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "Number of items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "List of history entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryEntry"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "delete": {
        "summary": "Clear History",
        "description": "Delete all browsing history for the user",
        "tags": ["History"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "204": {
            "description": "History cleared"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/search": {
      "get": {
        "summary": "Search",
        "description": "Search across bookmarks, history, and tabs",
        "tags": ["Search"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            },
            "description": "Search query"
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["bookmarks", "history", "tabs", "all"],
              "default": "all"
            },
            "description": "Type of content to search"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            },
            "description": "Maximum number of results"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResults"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Bookmark": {
        "type": "object",
        "properties": {
          "bookmarkId": {
            "type": "string",
            "description": "Unique bookmark identifier"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Bookmark URL"
          },
          "title": {
            "type": "string",
            "description": "Bookmark title"
          },
          "description": {
            "type": "string",
            "description": "Bookmark description"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Bookmark tags"
          },
          "folder": {
            "type": "string",
            "description": "Folder/category"
          },
          "faviconUrl": {
            "type": "string",
            "format": "uri",
            "description": "Favicon URL"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": ["bookmarkId", "url"]
      },
      "BookmarkInput": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Bookmark URL"
          },
          "title": {
            "type": "string",
            "description": "Bookmark title"
          },
          "description": {
            "type": "string",
            "description": "Bookmark description"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Bookmark tags"
          },
          "folder": {
            "type": "string",
            "description": "Folder/category"
          }
        },
        "required": ["url"]
      },
      "HistoryEntry": {
        "type": "object",
        "properties": {
          "historyId": {
            "type": "string",
            "description": "Unique history entry identifier"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Visited URL"
          },
          "title": {
            "type": "string",
            "description": "Page title"
          },
          "visitCount": {
            "type": "integer",
            "description": "Number of visits"
          },
          "lastVisit": {
            "type": "string",
            "format": "date-time",
            "description": "Last visit timestamp"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "First visit timestamp"
          }
        },
        "required": ["historyId", "url"]
      },
      "Tab": {
        "type": "object",
        "properties": {
          "tabId": {
            "type": "string",
            "description": "Unique tab identifier"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Current URL"
          },
          "title": {
            "type": "string",
            "description": "Page title"
          },
          "favicon": {
            "type": "string",
            "description": "Favicon data URL"
          },
          "loading": {
            "type": "boolean",
            "description": "Loading state"
          },
          "canGoBack": {
            "type": "boolean",
            "description": "Can navigate back"
          },
          "canGoForward": {
            "type": "boolean",
            "description": "Can navigate forward"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Tab creation timestamp"
          }
        },
        "required": ["tabId"]
      },
...
      "SearchResults": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string",
            "description": "Original search query"
          },
          "total": {
            "type": "integer",
            "description": "Total number of results"
          },
          "bookmarks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Bookmark"
            },
            "description": "Bookmark search results"
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HistoryEntry"
            },
            "description": "History search results"
          },
          "tabs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tab"
            },
            "description": "Tab search results"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "code": {
            "type": "string",
            "description": "Error code"
          },
          "requestId": {
            "type": "string",
            "description": "Request ID for tracking"
          }
        },
        "required": ["error"]
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}